requirements.txt
azure-functions==1.18.0
azure-ai-textanalytics==5.3.0
azure-identity==1.15.0
azure-core==1.29.5
azure-common==1.1.28